Fix right panel background extending beyond the truncation symbol
In side-by-side mode, if `background_color_extends_to_terminal_width`
is set, the left panel color is extended via spaces, but the right
one via an ANSI sequence which instructs the terminal emulator to
fill the background color rightwards.

However due to the padding after the right panel the color will
extend beyond a possible truncation symbol:

"TEXT#AB→#", not "TEXT#AB→ " with "#" being the green background.

If any line is truncated, the entire block is now padded with spaces
(if only selected lines used spaces the line would show a jagged
right edge)

Add enums `BgFillWidth` and `BgFillMethod` to better distinguish
if the background should be filled, and if so how (ANSI or Spaces).